requirements.txt
beautifulsoup4==4.10.0
disnake==2.3.0
PyYAML==6.0
requests==2.27.1
